+++ todotodo/models.py
@@ -1,4 +1,3 @@
-from tkinter import CASCADE
 from django.db import models
 from django.utils import timezone
 from django.contrib.auth.models import User
@@ -6,15 +5,6 @@ from django.contrib.postgres.fields import ArrayField
 from django.core.validators import MinValueValidator, MaxValueValidator
 
 # Create your models here.
-class Profile(models.Model):
-    name = models.CharField(max_length=256, default='')
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    friends = models.ManyToManyField('self', through='Friendship', related_name='friendship')
-    created_at = models.DateTimeField(default=timezone.now)
-
-    def __str__(self):
-        return self.name
-
 class Friendship(models.Model):
     friend1_id = models.ForeignKey(User, on_delete=models.CASCADE, related_name='friendship_sender') #sender
     friend2_id = models.ForeignKey(User, on_delete=models.CASCADE, related_name='friendship_receiver') #receiver
