+++ todotodo/templates/todotodo/login.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset='utf-8'>
+        <title>snulion</title>
+        {% load socialaccount %} {% providers_media_js %}
+    </head>
+    <body>
+        <h1>todotodo</h1>
+        <a href="{% provider_login_url 'kakao' method='oauth2' %}">
+            카카오 로그인
+        </a>
+    </body>
+</html>
