+++ todotodo/views.py
@@ -4,4 +4,7 @@ from .models import Persona
 # Create your views here.
 def index(request):
     personas = Persona.objects.all()
-    return render(request, 'todotodo/index.html', {'personas': personas})
+    return render(request, 'todotodo/index.html')
+
+def login(request):
+    return render(request, 'todotodo/login.html')
